api/requirements.txt
fastapi-cli==0.0.6
fastapi==0.115.6
sqlmodel==0.0.22
alembic==1.14.1
SQLAlchemy==2.0.36
uvicorn==0.32.1
pydantic==2.10.3
psycopg2-binary
numpy>=1.20.0
tensorflow>=2.13.0
scikit-learn>=1.0.0
joblib>=1.1.0
